+++ fabric/src/main/java/dev/nyon/magnetic/mixins/compat/rch/RightClickHarvestMixin.java
@@ -3,19 +3,24 @@ package dev.nyon.magnetic.mixins.compat.rch;
 import com.llamalad7.mixinextras.injector.v2.WrapWithCondition;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.llamalad7.mixinextras.sugar.Local;
 import dev.nyon.magnetic.utils.MixinHelper;
 import io.github.jamalam360.rightclickharvest.RightClickHarvest;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.BlockHitResult;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import java.util.List;
 import java.util.function.Consumer;
@@ -73,6 +78,27 @@ public class RightClickHarvestMixin {
         if (serverPlayer == null) return true;
 
         return MixinHelper.wrapWithConditionPlayerItemSingle(serverPlayer, stack);
+    }
+
+    @Inject(
+        method = "onBlockUse",
+        at = @At(
+            value = "INVOKE",
+            target = "Lio/github/jamalam360/rightclickharvest/RightClickHarvest;completeHarvest(Lnet/minecraft/world/level/Level;Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/entity/player/Player;Lnet/minecraft/world/InteractionHand;Lnet/minecraft/world/item/ItemStack;ZZLjava/lang/Runnable;)Lnet/minecraft/world/InteractionResult;",
+            ordinal = 1
+        )
+    )
+    private static void tagSurroundingBlocksWithPlayer(
+        Player player,
+        Level level,
+        InteractionHand hand,
+        BlockHitResult hitResult,
+        boolean initialCall,
+        CallbackInfoReturnable<InteractionResult> cir,
+        @Local(ordinal = 1) BlockPos breakPos
+    ) {
+        if (!(player instanceof ServerPlayer serverPlayer) || !(level instanceof ServerLevel serverLevel)) return;
+        MixinHelper.tagSurroundingBlocksWithPlayer(serverPlayer, breakPos, serverLevel);
     }
 
     @WrapWithCondition(